requirements.txt
flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
flask-socketio==5.3.5
eventlet==0.33.3
Flask-Session==0.5.0
